+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 - Fix autocorrection loop in `RSpec/ExampleWording` for insufficient example wording. ([@pirj])
+- Fix `RSpec/InferredSpecType` to skip removing `type: :routing` metadata. ([@pirj])
 - Add `named_only` style to `RSpec/NamedSubject`. ([@kuahyeow])
 - Fix a false positive for `RSpec/NoExpectationExample` when allowed pattern methods with arguments. ([@ydah])